integrating async downloads
